fix: ensure file size compliance (500-line limit)
- Reduced builder.py from 502 to 498 lines by combining typing imports
- Split main.py from 521 to 441 lines by extracting helpers and formatters
- Created cli_helpers.py (50 lines) for manager factory functions
- Created cli_formatters.py (53 lines) for display formatting functions
- All 41 tests passing ✓
- No functionality changed, pure refactoring for maintainability

Addresses code organization standards from project CLAUDE.md.
